Commands added and inventory changed and command handling moved
