+++ OOP_Sem1_Project/src/Command/Command.java
@@ -16,7 +16,7 @@
  * @author Michael Kolling and David J. Barnes
  * @version 2006.03.30
  */
-package oop_sem1_project;
+package Command;
 
 /**
  *
