+++ OOP_Sem1_Project/src/Command/CommandWord.java
@@ -1,4 +1,4 @@
-package oop_sem1_project;
+package Command;
 
 /**
  *
