+++ OOP_Sem1_Project/src/oop_sem1_project/Game.java
@@ -1,5 +1,8 @@
 package oop_sem1_project;
 
+import Command.CommandWord;
+import Command.Command;
+
 /**
  *
  *
