+++ OOP_Sem1_Project/src/oop_sem1_project/Room.java
@@ -1,8 +1,10 @@
 package oop_sem1_project;
 
+import java.util.ArrayList;
 import java.util.Set;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.List;
 
 /**
  *
@@ -20,21 +22,33 @@ public class Room {
      * A description of this Room.
      */
     private final String description;
-
+    /**
+     * 
+     */
+    private final String why;
     /**
      * Constructs a new Room Object.
      *
      * @param description A short description of this Room.
+     * @param why A brief description of how or why there is an accident or incident in a room
      */
-    public Room(String description) {
+    public Room(String description, String why) {
         this.description = description;
+        this.why = why;
+    }
+    
+    /**
+     * Constructs a new Room Object.
+     * @param description 
+     */
+    public Room(String description){
+        this(description, "Who knows");
     }
 
     /**
      * Add a new exit to this room.
-     *
-     * @param direction The relative direction in which the neighbor is located.
-     * @param neighbor  The neighboring room.
+     * @param direction The relative direction in which the neighbour is located.
+     * @param neighbor  The neighbour room.
      */
     public void setExit(String direction, Room neighbor) {
         this.exits.put(direction, neighbor);
@@ -65,7 +79,7 @@ public class Room {
         }
         return returnString;
     }
-
+     
     /**
      * Get an Exit in the given direction.
      *
