+++ OOP_Sem1_Project/src/oop_sem1_project/saftyPoint.java
@@ -1,0 +1,14 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package oop_sem1_project;
+
+/**
+ *
+ * @author Pinnacle F
+ */
+public class saftyPoint {
+    
+}
